Model rework

+++ backend/src/main/java/com/orbital/nusmaps/repository/EdgeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.orbital.nusmaps.model.Edge;
 
-public interface EdgeRepository extends JpaRepository<Edge, Integer> {
+public interface EdgeRepository extends JpaRepository<Edge, Long> {
 }
